Add comments and clean up some bits

+++ Sources/DistributedActorsDungeon/Mutex.swift
@@ -25,7 +25,7 @@ public final class Mutex {
   public init() {
     var attr: pthread_mutexattr_t = pthread_mutexattr_t()
     pthread_mutexattr_init(&attr)
-    pthread_mutexattr_settype(&attr, PTHREAD_MUTEX_RECURSIVE)
+    pthread_mutexattr_settype(&attr, Int32(PTHREAD_MUTEX_RECURSIVE))
 
     let error = pthread_mutex_init(&mutex, &attr)
     pthread_mutexattr_destroy(&attr)
